[frontend] update frontend when adding a user to a group (#6703)
